# docs/api/openapi.yaml
swagger: '2.0'
info:
  title: netobs unified API
  version: 2.5.6
paths:
...
  /api/v1/gpu-rca:
    get:
      description: 노드 하나의 GPU 유휴 dominant cause 와 cause 별 가중치 (한국어 설명 포함), 신뢰도, 원인
        후보 pod 랭킹, 근거 수치 (evidence), 한 줄 narrative 를 한 응답으로 합성한다. dominant cause 와
        가중치는 scope=node gpu-idle 결과를 재사용하고, 원인 후보는 이 노드를 victim 으로 하는 noisy-neighbor
        suspect pod (namespace-aware) 와 cross-node-interference suspect node 를 점수순으로
        집계한다. 신뢰도는 top1 과 top2 cause 격차이며 0.1 미만 백중세는 narrative 에 판정 유보로 적는다. evidence
        는 device 사용률과 SM active, 노드 재전송 rate 와 최대 RTT 에 더해 dominant cause 의 차원 맞춤
        수치 (memory 는 suspect pod 의 working_set/limit 과 노드 실측 사용률, thermal 은 온도와 slowdown
        여유, cpu 는 throttle 비율, 점수형 cause 는 노드 score) 를 cause 레지스트리 기반 2차 조회로 실어 narrative
        에 융합하고, dominant cause 별 인과 체인 문구를 덧붙인다. gpu 파라미터 (GPU UUID 또는 device index)
        로 evidence 의 GPU 수치를 device 로 좁힐 수 있고, 미등록 device 는 해당 수치가 생략된다.
      produces:
      - application/json
      tags:
      - gpu
      summary: 노드 GPU RCA 합성
      parameters:
      - type: string
        description: 대상 노드 (DNS-1123 형식)
        name: node
        in: query
        required: true
      - type: string
        description: 대상 device (GPU UUID 또는 device index, 생략 시 노드 전체)
        name: gpu
        in: query
      - type: integer
        description: 원인 후보 상위 N (1-50, 기본 10)
        name: limit
        in: query
      - type: string
        description: 평가 시점 (RFC3339 또는 unix seconds, 생략 시 현재)
        name: at
        in: query
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/internal_correlation_api.NodeGpuRcaResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
...
  /api/v1/rca:
    get:
      description: rca-summarizer 가 Alertmanager webhook 으로 생성한 alert 별 root cause
        analysis 요약을 프록시한다. alert 파라미터 생략 시 보관 중인 전체 요약 목록, 지정 시 해당 alert 의 최신 요약을
        돌려준다 (미보관 alert 는 404). 요약에는 지배 자원 차원과 최우선 의심 pod, 핵심 drop flow, 근거 메트릭 힌트,
        신뢰도 점수가 담긴다.
      produces:
      - application/json
      tags:
      - rca
      summary: RCA 요약 조회
      parameters:
      - type: string
        description: alertname (생략 시 전체 목록)
        name: alert
        in: query
      responses:
        '200':
          description: OK
          schema:
            type: object
            additionalProperties: true
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
        '502':
          description: Bad Gateway
          schema:
            $ref: '#/definitions/netobs_internal_apicommon.ErrorBody'
...
definitions:
...
  internal_correlation_api.GpuCauseWeight:
    type: object
    properties:
      cause:
        type: string
      chain:
        description: 'Chain 은 cause 의 단계형 인과 체인 문구다 (#303). 프론트가 narrative 파싱 없이 구조
          필드로

          렌더할 수 있게 gpu-rca 가 채우며, gpu-idle 응답에서는 Description 과 동일하게 생략된다.'
        type: string
      description:
        description: 'Description 은 cause 의 한국어 한 줄 설명이다 (#287 레지스트리). gpu-rca 가 채우며
          gpu-idle

          응답에서는 생략된다.'
        type: string
      weight:
        type: number
...
  internal_correlation_api.NodeGpuRcaResponse:
    type: object
    properties:
      causes:
        type: array
        items:
          $ref: '#/definitions/internal_correlation_api.GpuCauseWeight'
      confidence:
        description: 'Confidence 는 dominant 판정 신뢰도 (0-1). top1 과 top2 cause weight
          의 격차 (margin) 로,

          단일 cause 가 지배적일수록 1 에 가깝고 두 cause 가 백중이면 0 에 가깝다. GPUIdleDominantCause

          Ambiguous alert 의 격차 < 0.1 판정과 동일 축이다.'
        type: number
      dominant_cause:
        type: string
      evidence:
        description: Evidence 는 narrative 를 뒷받침하는 근거 수치다.
        allOf:
        - $ref: '#/definitions/internal_correlation_api.RcaEvidence'
      generated_at:
        type: string
      gpu:
        description: Gpu 는 device 스코프 조회 시 요청 gpu 파라미터 (GPU UUID 또는 device index)
          의 echo 다.
        type: string
      idle:
        description: Idle 은 node:gpu_idle:5m (0-1). 유휴 게이팅 미충족 시 cause 귀속이 비므로 함께
          노출한다.
        type: number
      narrative:
        type: string
      node:
        type: string
      suspects:
        type: array
        items:
          $ref: '#/definitions/internal_correlation_api.RcaSuspect'
...
  internal_correlation_api.RcaEvidence:
    type: object
    properties:
      cause_score:
        description: 'CauseScore 는 점수형 cause (pcie_saturation, dcgm_pcie_replay, nccl_collective_stall,

          host_compute_stall, cgroup_contention) 의 노드 score (0-1) 다.'
        type: number
      cpu_throttle_ratio:
        description: CpuThrottleRatio 는 dominant 가 cpu_throttle 일 때 이 노드 pod 의 최대
          throttle 비율 (0-1) 이다.
        type: number
      gpu_utilization_percent:
        description: GpuUtilizationPercent 는 device 사용률 (0-100) 이다. GPM 미지원 GPU 의
          SM active fallback 축.
        type: number
      max_srtt_seconds:
        description: 'MaxSrttSeconds 는 이 노드 연결들의 최대 smoothed RTT 다. latency-breakdown
          과 동일 소스

          (netobs_tcp_state_max_srtt_seconds) 로, p99 가 아닌 최대값이라 필드명도 max 로 적는다.'
        type: number
      node_memory_used_ratio:
        description: NodeMemoryUsedRatio 는 dominant 가 memory_pressure 일 때 노드 실측 메모리
          사용률 (0-1) 이다.
        type: number
      retrans_per_sec:
        description: RetransPerSec 는 이 노드의 TCP 재전송 rate (5m) 다.
        type: number
      slowdown_headroom_celsius:
        type: number
      sm_active_percent:
        description: SMActivePercent 는 gpm sm_occupancy 로, 데이터센터 GPU 에서만 채워진다.
        type: number
      suspect_memory_limit_ratio:
        description: 'SuspectMemoryLimitRatio 는 dominant 가 memory_pressure 일 때 최우선
          suspect pod 의

          working_set/limit 비율 (0-1) 이다. suspect 가 없으면 노드 pod 최대값으로 fallback 한다.'
        type: number
      temperature_celsius:
        description: 'TemperatureCelsius 와 SlowdownHeadroomCelsius 는 dominant 가 thermal
          일 때 노드 device 최고

          온도와 slowdown 임계까지의 여유다. 여유는 임계와 온도가 모두 수집됐을 때만 산출된다.'
        type: number
  internal_correlation_api.RcaSuspect:
    type: object
    properties:
      dimension:
        type: string
      issues:
        type: array
        items:
          type: string
      namespace:
        type: string
      node:
        type: string
      pod:
        type: string
      score:
        type: number
      source:
        type: string
...
  netobs_internal_apicommon.ErrorBody:
    type: object
    properties:
      error:
        $ref: '#/definitions/netobs_internal_apicommon.ErrorDetail'
  netobs_internal_apicommon.ErrorDetail:
    type: object
    properties:
      code:
        type: string
      message:
        type: string
...